Add `ProcessConfig` to fork or run in process
`ProcessConfig` is in charge of configuring whether a run will take
place inside the same JVM or if a new JVM will be forked. It takes care
of redirecting the output so that everything ends up in the logger.

Currently, the redirection of stdout and stderr prevents us from running
multiple processes without forking.